Адаптация списка компонентов под панель
